Add support for functionality extensions
Not auto-discovered yet
