+++ src/core/index.ts
@@ -1,4 +1,6 @@
 export * from './batches';
+export * from './config';
+export * from './extension';
 export * from './jobs';
 export * from './packages';
 export * from './plan';
